update: move passivation tests
